Moved handlers to separate package
