+++ handlers/api_handlers.go
@@ -1,4 +1,4 @@
-package middleware
+package handlers
 
 import (
 	"net/http"
