+++ handlers/status_handler.go
@@ -1,4 +1,4 @@
-package middleware
+package handlers
 
 import (
 	"net/http"
